Add blurbs to Blog and News
The code for the blog is a horror show at the moment. Lots of elements tightly tangled in places where they really shouldn't be. Definitely shouldn't be merged in its current state, but some creative gymnastics are going to be needed to circumvent type limitations on the blog plugin.

News, thankfully, is pretty much final.